# projects/xobusy/sunrint-decon/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/register":{"post":{"summary":"Register","operationId":"register_register_post","parameters":[{"name":"name","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"}},{"name":"password","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Password"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/school/search":{"get":{"summary":"School Search","operationId":"school_search_school_search_get","parameters":[{"name":"name","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/School"},"title":"Response School Search School Search Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/school/review":{"get":{"summary":"School Review","operationId":"school_review_school_review_get","parameters":[{"name":"code","in":"query","required":true,"schema":{"type":"integer","title":"Code"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Review"},"title":"Response School Review School Review Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"School Review Post","operationId":"school_review_post_school_review_post","parameters":[{"name":"code","in":"header","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Code"}},{"name":"username","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Username"}},{"name":"password","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Password"}},{"name":"review","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Review"}},{"name":"stars","in":"header","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Stars"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Review":{"properties":{"username":{"type":"string","title":"Username"},"review":{"type":"string","title":"Review"},"star":{"type":"number","title":"Star"}},"type":"object","required":["username","review","star"],"title":"Review"},"School":{"properties":{"name":{"type":"string","title":"Name"},"kind":{"type":"string","title":"Kind"},"code":{"type":"integer","title":"Code"},"homepage":{"type":"string","title":"Homepage"},"address":{"type":"string","title":"Address"},"callnumber":{"type":"string","title":"Callnumber"},"created":{"type":"string","title":"Created"}},"type":"object","required":["name","kind","code","homepage","address","callnumber","created"],"title":"School"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}